Fix CI failures: clippy warning and CUDA panic handling
- Remove unnecessary u32 cast in cuda_packed.rs:429
- Wrap is_cuda_available() and cuda_device_count() with catch_unwind
  to handle cudarc panics when CUDA libraries are not installed

The cudarc library panics at runtime when CUDA libraries are not
found on the system (e.g., in CI without GPU). By catching these
panics, the Backend::Auto detection now gracefully falls back to
CPU when CUDA is unavailable instead of crashing.

+++ crates/ringkernel-wavesim/src/simulation/cuda_packed.rs
@@ -426,7 +426,7 @@ impl CudaPackedBackend {
                 RingKernelError::BackendError("apply_boundary_conditions not found".into())
             })?;
 
-        let max_edge_threads = std::cmp::max(self.tiles_x, self.tiles_y) as u32 * self.tile_size;
+        let max_edge_threads = std::cmp::max(self.tiles_x, self.tiles_y) * self.tile_size;
 
         let boundary_cfg = cudarc::driver::LaunchConfig {
             grid_dim: (4, 1, 1), // 4 edges
